refactor: move simulation output builder module
